feat: change addAlert to alert
