+++ package.json
@@ -5,6 +5,8 @@
   "scripts": {},
   "private": true,
   "dependencies": {
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
     "@swc/helpers": "~0.5.0",
     "alertx": "^0.0.7",
     "classnames": "^2.3.2",
